Move RSVP controller test into `controller` dir and fix spelling mistake in dir name
